requirements.txt
Flask
flask-cors
groq
gunicorn
python-dotenv
